conformance/buffimgs/cl_commandQueue_enqueueReadBufferRect.html

<!--
 Copyright (C) 2013 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following condition
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>WebCLCommandQueue enqueueReadBufferRect conformance test.</title>
    <link rel="stylesheet" href="../../resources/js-test-style.css"/>
    <script src="../../resources/js-test-pre.js"></script>
    <script src="../../resources/cl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>

try {
    debug("WebCLCommandQueue enqueueReadBufferRect test case.\n");
    debug("( WebCLBuffer buffer, CLboolean blockingRead, sequence<CLuint> bufferOrigin, sequence<CLuint> hostOrigin, sequence<CLuint> region, CLuint bufferRowPitch, CLuint bufferSlicePitch, CLuint hostRowPitch, CLuint hostSlicePitch, ArrayBufferView hostPtr, optional sequence<WebCLEvent>? eventWaitList, optional WebCLEvent? event)\n ");

    var bufferRowPitch = 0;
    var bufferSlicePitch = 0;
    var hostRowPitch = 0;
    var hostSlicePitch = 0;
    var wtu = WebCLTestUtils;
    var webCLContext = wtu.createContext();
    var webCLCommandQueue = wtu.createCommandQueue(webCLContext);
    var webCLBufferOrigin = [0, 0, 0];
    var hostOrigin = [0, 0, 0];
    var region = [10, 10, 1];
    var hostPtr = new Float32Array(BUFFER_SIZE);

    // Testing with buffer as MEM_READ_WRITE memory object.
    var webCLBufferReadWrite = wtu.createBuffer(webCLContext, webcl.MEM_READ_WRITE, BUFFER_SIZE);
    shouldBeUndefined("webCLCommandQueue.enqueueReadBufferRect(webCLBufferReadWrite, true, webCLBufferOrigin, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, hostPtr);");
    shouldBeUndefined("webCLCommandQueue.enqueueReadBufferRect(webCLBufferReadWrite, false, webCLBufferOrigin, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, hostPtr);");

    // Testing with buffer as MEM_READ_ONLY memory object.
    var webCLBufferRead = wtu.createBuffer(webCLContext, webcl.MEM_READ_ONLY, BUFFER_SIZE);
    shouldBeUndefined("webCLCommandQueue.enqueueReadBufferRect(webCLBufferRead, true, webCLBufferOrigin, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, hostPtr);");
    shouldBeUndefined("webCLCommandQueue.enqueueReadBufferRect(webCLBufferRead, true, webCLBufferOrigin, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, hostPtr);");

    // Testing with buffer as MEM_WRITE_ONLY memory object.
    var webCLBufferWrite = wtu.createBuffer(webCLContext, webcl.MEM_WRITE_ONLY, BUFFER_SIZE);
    shouldBeUndefined("webCLCommandQueue.enqueueReadBufferRect(webCLBufferWrite, true, webCLBufferOrigin, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, hostPtr);");
    shouldBeUndefined("webCLCommandQueue.enqueueReadBufferRect(webCLBufferWrite, true, webCLBufferOrigin, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, hostPtr);");

    var webCLBuffer = wtu.createBuffer(webCLContext, webcl.MEM_READ_WRITE, BUFFER_SIZE);

    // Testing for values of buffer Origin.
    var webCLBufferOriginInvalid = [];
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadBufferRect(webCLBuffer, true, webCLBufferOriginInvalid, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, hostPtr);", "INVALID_VALUE");
    webCLBufferOriginInvalid = [BUFFER_SIZE + 1, BUFFER_SIZE + 1, 0];
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadBufferRect(webCLBuffer, true, webCLBufferOriginInvalid, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, hostPtr);", "INVALID_VALUE");
    webCLBufferOriginInvalid = new Int32Array(1);
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadBufferRect(webCLBuffer, true, webCLBufferOriginInvalid, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, hostPtr);", "INVALID_VALUE");
    webCLBufferOriginInvalid = new Int32Array(2);
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadBufferRect(webCLBuffer, true, webCLBufferOriginInvalid, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, hostPtr);", "INVALID_VALUE");

    // Testing for values of host Origin.
    var hostOriginInvalid = [];
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadBufferRect(webCLBuffer, true, webCLBufferOrigin, hostOriginInvalid, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, hostPtr);", "INVALID_VALUE");
    hostOriginInvalid = [BUFFER_SIZE + 1, BUFFER_SIZE + 1, 0];
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadBufferRect(webCLBuffer, true, webCLBufferOrigin, hostOriginInvalid, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, hostPtr);", "INVALID_VALUE");
    hostOriginInvalid = new Int32Array(1);
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadBufferRect(webCLBuffer, true, webCLBufferOrigin, hostOriginInvalid, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, hostPtr);", "INVALID_VALUE");
    hostOriginInvalid = new Int32Array(2);
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadBufferRect(webCLBuffer, true, webCLBufferOrigin, hostOriginInvalid, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, hostPtr);", "INVALID_VALUE");

    // Testing for values of rectangular region being read.
    var regionInvalid = [];
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadBufferRect(webCLBuffer, true, webCLBufferOrigin, hostOrigin, regionInvalid, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, hostPtr);", "INVALID_VALUE");
    regionInvalid = [BUFFER_SIZE + 1, BUFFER_SIZE + 1];
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadBufferRect(webCLBuffer, true, webCLBufferOrigin, hostOrigin, regionInvalid, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, hostPtr);", "INVALID_VALUE");
    regionInvalid = new Int32Array(1);
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadBufferRect(webCLBuffer, true, webCLBufferOrigin, hostOrigin, regionInvalid, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, hostPtr);", "INVALID_VALUE");
    regionInvalid = new Int32Array(2);
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadBufferRect(webCLBuffer, true, webCLBufferOrigin, hostOrigin, regionInvalid, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, hostPtr);", "INVALID_VALUE");

    // Testing for values of bufferRowPitch.
    var bufferRowPitchValid = WIDTH_STEP;
    shouldBeUndefined("webCLCommandQueue.enqueueReadBufferRect(webCLBuffer, true, webCLBufferOrigin, hostOrigin, region, bufferRowPitchValid, bufferSlicePitch, hostRowPitch, hostSlicePitch, hostPtr);");
    var bufferRowPitchInvalid = BUFFER_SIZE;
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadBufferRect(webCLBuffer, true, webCLBufferOrigin, hostOrigin, region, bufferRowPitchInvalid, bufferSlicePitch, hostRowPitch, hostSlicePitch, hostPtr);", "INVALID_VALUE");

    // Testing for values of bufferSlicePitch.
    var bufferSlicePitchValid = WIDTH_STEP;
    shouldBeUndefined("webCLCommandQueue.enqueueReadBufferRect(webCLBuffer, true, webCLBufferOrigin, hostOrigin, region, bufferRowPitch, bufferSlicePitchValid, hostRowPitch, hostSlicePitch, hostPtr);");
    var bufferSlicePitchInvalid = (bufferRowPitchValid * region[1]) - 1;
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadBufferRect(webCLBuffer, true, webCLBufferOrigin, hostOrigin, region, bufferRowPitch, bufferSlicePitchInvalid, hostRowPitch, hostSlicePitch, hostPtr);", "INVALID_VALUE");

    // Testing for values of hostRowPitchInvalid.
    var hostRowPitchValid = WIDTH_STEP;
    shouldBeUndefined("webCLCommandQueue.enqueueReadBufferRect(webCLBuffer, true, webCLBufferOrigin, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitchValid, hostSlicePitch, hostPtr);");
    var hostRowPitchInvalid = BUFFER_SIZE;
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadBufferRect(webCLBuffer, true, webCLBufferOrigin, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitchInvalid, hostSlicePitch, hostPtr);", "INVALID_VALUE");

    // Testing for values of hostSlicePitch.
    var hostSlicePitchValid = WIDTH_STEP;
    shouldBeUndefined("webCLCommandQueue.enqueueReadBufferRect(webCLBuffer, true, webCLBufferOrigin, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitchValid, hostPtr);");
    var hostSlicePitchInvalid = (hostRowPitchValid * region[1]) - 1;
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadBufferRect(webCLBuffer, true, webCLBufferOrigin, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitchInvalid, hostPtr);", "INVALID_VALUE");

    // Testing for values of hostPtr.
    var hostPtrInvalid = new Float32Array(BUFFER_SIZE - 1);
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadBufferRect(webCLBuffer, true, webCLBufferOrigin, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, hostPtrInvalid);", "INVALID_VALUE");

    // Testing for exception "INVALID_MEM_OBJECT".
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadBufferRect(invalid_object, true, webCLBufferOrigin, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, hostPtr);", "INVALID_MEM_OBJECT");

    // Testing for exception "INVALID_CONTEXT".
    var webCLContextNew = wtu.createContext();
    var webCLCommandQueueNew = wtu.createCommandQueue(webCLContextNew);
    var webCLBufferNew = wtu.createBuffer(webCLContextNew, webcl.MEM_READ_WRITE, BUFFER_SIZE);
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadBufferRect(webCLBufferNew, true, webCLBufferOrigin, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, hostPtr);", "INVALID_CONTEXT");
    shouldThrowExceptionName("webCLCommandQueueNew.enqueueReadBufferRect(webCLBuffer, true, webCLBufferOrigin, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, hostPtr);", "INVALID_CONTEXT");

    var webCLEvent = wtu.createEvent(webCLContext);
    var webCLEventNew = wtu.createEvent(webCLContextNew);
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadBufferRect(webCLBuffer, true, webCLBufferOrigin, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, hostPtr, [webCLEventNew], null);", "INVALID_CONTEXT");
    shouldThrowExceptionName("webCLCommandQueueNew.enqueueReadBufferRect(webCLBuffer, true, webCLBufferOrigin, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, hostPtr, [webCLEvent], null);", "INVALID_CONTEXT");
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadBufferRect(webCLBuffer, true, webCLBufferOrigin, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, hostPtr, [], webCLEventNew);", "INVALID_CONTEXT");
    shouldThrowExceptionName("webCLCommandQueueNew.enqueueReadBufferRect(webCLBuffer, true, webCLBufferOrigin, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, hostPtr, [], webCLEvent);", "INVALID_CONTEXT");

    wtu.release(webCLEventNew);
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadBufferRect(webCLBuffer, true, webCLBufferOrigin, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, hostPtr, [webCLEventNew], webCLEvent);", "INVALID_EVENT_WAIT_LIST");

    webCLEvent.setUserEventStatus(-1);
    shouldThrowExceptionName("webCLCommandQueue.enqueueReadBufferRect(webCLBuffer, true, webCLBufferOrigin, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, hostPtr, [webCLEvent]);", "EXEC_STATUS_ERROR_FOR_EVENTS_IN_WAIT_LIST");

} catch (e) {
    testFailed(e.description);
}
successfullyParsed = true;
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
